refactor(agent-tool): split orchestration lifecycle
